fix hard code data and lost user data via session

+++ src/components/MUIComponent/RecentlyVisited.jsx
@@ -63,21 +63,22 @@ const RecentlyVisited = () => {
   const archiveNote = async (noteId) => {
     console.log("Archive note " + noteId);
     try {
-      const response = await fetch(
-        `http://localhost:8080/api/v1/note/notes?noteId=${noteId}`,
-        {
-          method: "DELETE",
-          credentials: "include",
-          headers: {
-            "Content-Type": "application/json",
-          },
-        }
-      );
-      const jsonData = await response.json();
-      console.log(jsonData.data);
-      if (response.ok) {
-        console.log(jsonData.message);
-        setNotes((prevNotes) => prevNotes.filter((note) => note.id !== noteId));
+      const noteIndex = noteIds.indexOf(noteId);
+      if (noteIndex !== -1) {
+        const deletedNote = noteTitles[noteIndex];
+        const response = await fetch(
+          `http://localhost:8080/api/v1/note/notes?id=${noteId}`,
+          {
+            method: "DELETE",
+            credentials: "include",
+          }
+        );
+        const data = await response.json();
+        console.log(data.message);
+        setNoteTitles((prevTitles) =>
+          prevTitles.filter((title) => title !== deletedNote)
+        );
+        setNoteIds((prevIds) => prevIds.filter((id) => id !== noteId));
       }
     } catch (error) {
       console.log("Archive note error: " + error);
@@ -86,7 +87,7 @@ const RecentlyVisited = () => {
 
   const navigate = useNavigate();
   const navToNoteScreen = (noteId) => {
-    navigate(`/note/${noteId}`);
+    navigate(`/note?id=${noteId}`);
   };
 
   return (
